Optimize GUI locks for tx generated type lookups
This optimizes how the GUI determines the "mined" type of a coinstake
transaction. It reorganizes the lookups to avoid the cs_main locks if
the node is busy with another operation. This prevents GUI stalls for
routine updates of the overview and history pages from core state. It
prevents the long-running polls load on the voting page from bricking
the GUI thread.
